sync: move processTask logic into walk
Walk was adding all the files it finds to the task channel, resulting
with flooding the channel with task that are already downloaded
alongside with new tasks as well. It was hard for consumers to find
failed tasks with this behaviour.

Walk is now pushes *new* tasks only, so that there are room for failed
tasks.